backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "databricks-sdk[openai]",
    "pydantic",
    "starlette>=0.46.2",
    "click>=8.2.1",
    "mlflow>=3.0.0rc3",
    "mlflow-skinny>=3.0.0rc3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
